Read a SNES controller.

+++ LCDBusReader/src/ShiftRegister.h
@@ -42,7 +42,7 @@ void init_shift(S595OutPins* psPins);
 void shift_out(S595OutPins* psPins, uint16_t uiData);
 
 void init_shift(S165InPins* psPins);
-uint shift_in(S165InPins* psPins);
+uint16_t shift_in(S165InPins* psPins);
 
 
 #endif // __SHIFT_REGISTER_H__
